--- tests/fixtures/config-initializer/lib/doubleQuotes.js
@@ -1,1 +1,0 @@
-var foo = "doubleQuotes";
