--- tests/fixtures/config-initializer/lib/no-semi.js
@@ -1,1 +1,0 @@
-var name = "ESLint"
