--- tests/fixtures/config-initializer/new-es-features/new-es-features.js
@@ -1,3 +1,0 @@
-async function fn() {
-    await Promise.resolve();
-}
