--- tests/fixtures/config-initializer/parse-error/parse-error.js
@@ -1,1 +1,0 @@
-+;
