--- tests/fixtures/config-initializer/singleQuotes.js
@@ -1,1 +1,0 @@
-var foo = 'singleQuotes';
