--- tests/fixtures/config-initializer/tests/console-log.js
@@ -1,1 +1,0 @@
-console.log("I'm a log");
