--- tests/fixtures/config-initializer/tests/doubleQuotes.js
@@ -1,1 +1,0 @@
-var foo = "doubleQuotes";
